restructure: move parse functions
