+++ src/index.ts
@@ -1,14 +1,15 @@
 export { FefeError } from './errors'
 
 export { Validator } from './validate'
-export { array, ArrayOptions } from './array'
+export { array } from './array'
 export { boolean } from './boolean'
-export { date, DateOptions } from './date'
+export { date } from './date'
 export { _enum as enum } from './enum'
-export { number, NumberOptions } from './number'
-export { object, ObjectDefinition, ObjectDefinitionValue, ObjectOptions } from './object'
-export { string, StringOptions } from './string'
+export { number } from './number'
+export { object } from './object'
+export { parseBoolean } from './parse-boolean'
+export { parseDate } from './parse-date'
+export { parseJson } from './parse-json'
+export { parseNumber } from './parse-number'
+export { string } from './string'
 export { union } from './union'
-
-import * as transform from './transform/transform'
-export { transform }
