+++ src/parse-boolean.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
 
-import { FefeError } from '../errors'
+import { FefeError } from './errors'
 import { parseBoolean } from './parse-boolean'
 
 describe('parseBoolean()', () => {
