+++ src/parse-date.ts
@@ -1,9 +1,10 @@
-import { FefeError } from '../errors'
+import { FefeError } from './errors'
 
 const isoDateRegex = /\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d\.\d+([+-][0-2]\d:[0-5]\d|Z)/
 
 export function parseDate ({ iso = false }: { iso?: boolean } = {}) {
-  return (value: string) => {
+  return (value: unknown) => {
+    if (typeof value !== 'string') throw new FefeError(value, 'Not a string.')
     if (iso && !isoDateRegex.test(value)) throw new FefeError(value, 'Not an ISO 8601 date string.')
     const time = Date.parse(value)
     if (Number.isNaN(time)) throw new FefeError(value, 'Not a date.')
