+++ src/parse-json.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
 
-import { FefeError } from '../errors'
+import { FefeError } from './errors'
 import { parseJson } from './parse-json'
 
 describe('parseJson()', () => {
