+++ src/parse-number.ts
@@ -1,7 +1,8 @@
-import { FefeError } from '../errors'
+import { FefeError } from './errors'
 
 export function parseNumber () {
-  return (value: string) => {
+  return (value: unknown) => {
+    if (typeof value !== 'string') throw new FefeError(value, 'Not a string.')
     const num = parseFloat(value)
     if (Number.isNaN(num)) throw new FefeError(value, 'Not a number.')
     return num
